Seperated Common, Web, and Node class
