+++ index.ts
@@ -1,2 +1,2 @@
-export * as koi_tools from "./lib/tools";
-export * as koi_node from "./lib/node";
+export * as koi_common from "./src/common";
+export * as koi_web from "./src/web";
