+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "koi-tools.js",
   "scripts": {
-    "build": "webpack",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -22,7 +22,6 @@
     "axios": "^0.21.1",
     "human-crypto-keys": "^0.1.4",
     "libp2p-crypto": "^0.19.4",
-    "nedb-promises": "^4.1.3",
     "node-polyfill-webpack-plugin": "^1.1.1",
     "smartweave": "^0.4.29"
   },
